MAINT: move spherical_bessel_smooth
